+++ docs/changelog.md
@@ -10,6 +10,7 @@
 * Added ignorable cells to Otter Assign
 * Added `autograder_files` configuration for Otter Assign
 * Added passdown of assignment configurations to Otter Generate from Otter Assign
+* Fixed whitespace bug in Assign solution parsing
 
 **v1.1.6:**